module the functional processing unit of C++
