+++ jcppkernel/realtime_subprocess.py
@@ -1,0 +1,50 @@
+from queue import Queue
+from threading import Thread
+import subprocess
+
+class RealTimeSubprocess(subprocess.Popen):
+    inputRequest = "<inputRequest>"
+
+    def __init__(self, cmd, write_to_stdout, write_to_stderr, read_from_stdin):
+        self._write_to_stdout = write_to_stdout
+        self._write_to_stderr = write_to_stderr
+        self._read_from_stdin = read_from_stdin
+        super().__init__(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE,bufsize=0)
+        self._stdout_queue = Queue()
+        self._stdout_thread = Thread(target=RealTimeSubprocess._enqueue_output, args=(self.stdout, self._stdout_queue))
+        self._stdout_thread.daemon = True
+        self._stdout_thread.start()
+        self._stderr_queue = Queue()
+        self._stderr_thread = Thread(target=RealTimeSubprocess._enqueue_output, args=(self.stderr, self._stderr_queue))
+        self._stderr_thread.daemon = True
+        self._stderr_thread.start()
+
+    @staticmethod
+    def _enqueue_output(stream, queue):
+        for line in iter(lambda: stream.read(65535), b""): queue.put(line)
+        stream.close()
+
+    def write_contents(self):
+        def read_all_from_queue(queue):
+            res = b""
+            size = queue.qsize()
+            while size != 0:
+                res += queue.get_nowait()
+                size -= 1
+            return res
+
+        stderr_contents = read_all_from_queue(self._stderr_queue)
+        if stderr_contents: self._write_to_stderr(stderr_contents.decode())
+
+        stdout_contents = read_all_from_queue(self._stdout_queue)
+        if stdout_contents:
+            contents = stdout_contents.decode()
+            start = contents.find(self.__class__.inputRequest)
+            if start >= 0:
+                contents = contents.replace(self.__class__.inputRequest, "")
+                if len(contents) > 0: self._write_to_stdout(contents)
+                readLine = ""
+                while len(readLine) == 0: readLine = self._read_from_stdin()
+                readLine += "\n"
+                self.stdin.write(readLine.encode())
+            else: self._write_to_stdout(contents)
